Move WrappedFuture to top-level class
