+++ instrumentation/jaxrs-client/jaxrs-client-2.0/jaxrs-client-2.0-jersey-2.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/jaxrsclient/v2_0/JerseyClientInstrumentationModule.java
@@ -19,10 +19,7 @@ import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Future;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.TimeoutException;
 import net.bytebuddy.asm.Advice;
 import net.bytebuddy.description.method.MethodDescription;
 import net.bytebuddy.description.type.TypeDescription;
@@ -90,59 +87,6 @@ public class JerseyClientInstrumentationModule extends InstrumentationModule {
         @Advice.Return(readOnly = false) Future<?> future) {
       if (!(future instanceof WrappedFuture)) {
         future = new WrappedFuture<>(future, context);
-      }
-    }
-  }
-
-  public static class WrappedFuture<T> implements Future<T> {
-
-    private final Future<T> wrapped;
-    private final ClientRequest context;
-
-    public WrappedFuture(Future<T> wrapped, ClientRequest context) {
-      this.wrapped = wrapped;
-      this.context = context;
-    }
-
-    @Override
-    public boolean cancel(boolean mayInterruptIfRunning) {
-      return wrapped.cancel(mayInterruptIfRunning);
-    }
-
-    @Override
-    public boolean isCancelled() {
-      return wrapped.isCancelled();
-    }
-
-    @Override
-    public boolean isDone() {
-      return wrapped.isDone();
-    }
-
-    @Override
-    public T get() throws InterruptedException, ExecutionException {
-      try {
-        return wrapped.get();
-      } catch (ExecutionException e) {
-        Object prop = context.getProperty(ClientTracingFilter.SPAN_PROPERTY_NAME);
-        if (prop instanceof Span) {
-          tracer().endExceptionally((Span) prop, e.getCause());
-        }
-        throw e;
-      }
-    }
-
-    @Override
-    public T get(long timeout, TimeUnit unit)
-        throws InterruptedException, ExecutionException, TimeoutException {
-      try {
-        return wrapped.get(timeout, unit);
-      } catch (ExecutionException e) {
-        Object prop = context.getProperty(ClientTracingFilter.SPAN_PROPERTY_NAME);
-        if (prop instanceof Span) {
-          tracer().endExceptionally((Span) prop, e.getCause());
-        }
-        throw e;
       }
     }
   }
